requirements.txt
Django
django-nested-inline
django-bootstrap-datepicker-plus
